# docs/openapi-petugas.yaml
openapi: 3.1.0
info:
  title: JAGAPADI API — Kontrak Role Petugas
  version: 1.0.0
  description: >-
    Kontrak machine-readable operasi backend v1 yang digunakan Petugas. Semua
    resource laporan otomatis scoped ke subjek JWT. user_id, role, nomor laporan,
    verifier, dan status administratif dari client diabaikan atau ditolak.
servers:
  - url: http://localhost:8080/api/v1
    description: Backend v1 lokal
  - url: https://example.invalid/api/v1
    description: Ganti dengan origin deployment
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login Petugas
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {$ref: '#/components/responses/LoginSuccess'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 50}
        password: {type: string, format: password, minLength: 1}
    User:
      type: object
      required: [id, username, role]
      properties:
        id: {type: integer}
        username: {type: string}
        nama_lengkap: {type: string}
        role: {type: string, const: petugas}
        must_change_password: {type: boolean}
...
    Error:
      type: object
      required: [success, error, message]
      properties:
        success: {type: boolean, const: false}
        error: {type: string}
        message: {type: string}
        errors: {type: object, additionalProperties: true}
...
  responses:
...
    LoginSuccess:
      description: Login/token berhasil
      content:
        application/json:
          schema:
            type: object
            properties:
              success: {type: boolean, const: true}
              data:
                type: object
                properties:
                  access_token: {type: string}
                  token_type: {type: string, const: Bearer}
                  expires_in: {type: integer}
                  user: {$ref: '#/components/schemas/User'}
    Unauthenticated:
      description: Token tidak ada/tidak valid/expired/revoked
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
    RateLimited:
      description: Rate limit terlampaui
      headers:
        Retry-After: {schema: {type: integer}}
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}